move ticker to utils
