+++ utils/ticker.go
@@ -1,13 +1,16 @@
 // Copyright (C) 2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package database
+package utils
 
 import (
 	"sync"
 	"time"
 )
 
+// Ticker is a timer that can be subscribed to. When the timer ticks,
+// all subscribers will receive a signal on the channel they were given
+// when subscribing.
 type Ticker struct {
 	interval      time.Duration
 	subscriptions []chan struct{}
